Split up into modules:
* prs: Model, Geometry, Control, Result, run, read functions, equivalent_phases
* frs: interaction with fraysum
