+++ pyraysum/tests/test_0_imports.py
@@ -6,7 +6,8 @@ def test_imports():
     import numpy
     import matplotlib
     import pyraysum
-    from pyraysum import prs, Model, Result
+    from pyraysum import prs, frs
+    from pyraysum import Model, Geometry, Control, Result, run
     import fraysum
     assert True
     return
